split orders.tsx file & add toast
