+++ src/interfaces/common.interface.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from 'react'
+
+export interface IToastFunctionOptions {
+  severity?: 'info' | 'success' | 'warn' | 'error' | 'secondary' | 'contrast' | undefined
+  summary?: ReactNode
+  detail?: ReactNode
+  life?: number
+}
