--- src/pages/Orders.tsx
@@ -1,509 +1,0 @@
-import { useEffect, useMemo, useRef, useState } from 'react'
-import { Link, useLocation, useNavigate } from 'react-router-dom'
-
-import { Button } from 'primereact/button'
-import { Column } from 'primereact/column'
-import { ConfirmDialog, confirmDialog } from 'primereact/confirmdialog'
-import { DataTable } from 'primereact/datatable'
-import { Dropdown } from 'primereact/dropdown'
-import { InputText } from 'primereact/inputtext'
-import { OverlayPanel } from 'primereact/overlaypanel'
-import { Paginator } from 'primereact/paginator'
-
-import {
-  ConvertDateString,
-  ConvertTimeString,
-  filterEnumMapping,
-  getOrderStatusByEnum,
-  icons,
-  OrderStatus,
-  PaymentMethod,
-  PaymentStatus,
-  sortByEnumMapping,
-  FilterOptions,
-  SortOptions,
-  getPaymentStatusByEnum
-} from '@/utils'
-import { orderApi } from '@/apis'
-import { useApi } from '@/hooks'
-import { IOrderReturn } from '@/interfaces'
-
-type PaginationType = {
-  totalPages?: number
-  totalOrders?: number
-  currentPage?: number
-  limit?: number
-}
-type SortType = {
-  value: string
-  label: string
-  icon: JSX.Element
-}
-export function Orders() {
-  const location = useLocation()
-  const dropdownControlRef = useRef<OverlayPanel>(null)
-  const navigate = useNavigate()
-
-  const queryParams = useMemo(() => new URLSearchParams(location.search), [location.search])
-  const queryParamPage = parseInt(queryParams.get('page') || '1')
-  const queryParamLimit = parseInt(queryParams.get('limit') || '5')
-  const queryParamSortBy = sortByEnumMapping(queryParams.get('sortBy') ?? SortOptions.DateDecrease)
-  const queryParamFilter = filterEnumMapping(queryParams.get('filter') ?? FilterOptions.None)
-
-  const { loading: callOrderApiLoading, callApi: callOrderApi } = useApi<void>()
-
-  const [orders, setOrders] = useState<IOrderReturn[]>([])
-  const [selectedOrder, setSelectedOrder] = useState<IOrderReturn>()
-
-  const [pagination, setPagination] = useState<PaginationType>({})
-  const [inputKeyword, setInputKeyword] = useState<string>(queryParams.get('keyword') ?? '')
-  const [currentKeyword, setCurrentKeyword] = useState<string>(queryParams.get('keyword') ?? '')
-  const [currentSortBy, setCurrentSortBy] = useState<SortOptions>(queryParamSortBy)
-  const [currentFilter, setCurrentFilter] = useState<FilterOptions>(queryParamFilter)
-
-  const optionTemplate = (option: SortType) => {
-    return (
-      <div className='flex align-items-center'>
-        <div className='flex items-center gap-2'>
-          {option.label}
-          {option.icon}
-        </div>
-      </div>
-    )
-  }
-  const sortOptionTemplate = (option: SortType) => {
-    return optionTemplate(option)
-  }
-  const selectedOptionTemplate = (option: SortType) => {
-    return optionTemplate(option)
-  }
-  const sortOptions = [
-    {
-      value: SortOptions.DateDecrease,
-      label: 'Date: ',
-      icon: icons.downArrow
-    },
-    {
-      value: SortOptions.DateIncrease,
-      label: 'Date: ',
-      icon: icons.upArrow
-    },
-    {
-      value: SortOptions.PriceDecrease,
-      label: 'Price: ',
-      icon: icons.downArrow
-    },
-    {
-      value: SortOptions.PriceIncrease,
-      label: 'Price: ',
-      icon: icons.upArrow
-    }
-  ]
-  const filterOptions = [
-    {
-      value: FilterOptions.Delivered,
-      label: FilterOptions.Delivered
-    },
-    {
-      value: FilterOptions.Delivering,
-      label: FilterOptions.Delivering
-    },
-    {
-      value: FilterOptions.Confirmed,
-      label: FilterOptions.Confirmed
-    },
-
-    {
-      value: FilterOptions.Pending,
-      label: FilterOptions.Pending
-    },
-    {
-      value: FilterOptions.Canceled,
-      label: FilterOptions.Canceled
-    },
-    {
-      value: FilterOptions.None,
-      label: FilterOptions.None
-    }
-  ]
-  const deleteConfirm = () => {
-    confirmDialog({
-      message: 'Are you sure you want to delete this order?',
-      header: 'Delete',
-      icon: <span className='text-red-500 text-2xl'>{icons.warning}</span>,
-      defaultFocus: 'accept',
-      accept: () => handleDeleteOrder(),
-      reject: () => {}
-    })
-  }
-  const handleDeleteOrder = () => {
-    selectedOrder &&
-      callOrderApi(async () => {
-        const data = await orderApi.deleteOrder(selectedOrder.id)
-        if (data) {
-          fetchOrderWithCurrentParams()
-        } else {
-          // toast
-        }
-      })
-  }
-  const handleRestoreOrder = () => {
-    selectedOrder &&
-      callOrderApi(async () => {
-        const data = await orderApi.restoreOrder(selectedOrder.id)
-        if (data) {
-          fetchOrderWithCurrentParams()
-        } else {
-          // toast
-        }
-      })
-  }
-  const getNavigateParams = (limit: number = 5, keyword?: string, sortBy?: SortOptions, filter?: FilterOptions) => {
-    return {
-      sortBy: sortBy ? `&sortBy=${sortBy}` : '',
-      filter: filter ? `&filter=${filter}` : '',
-      limit: limit ? `&limit=${limit}` : '',
-      keyword: keyword ? `&keyword=${keyword}` : ''
-    }
-  }
-  const fetchOrders = (
-    page: number = 1,
-    limit: number = 5,
-    keyword?: string,
-    sortBy?: SortOptions,
-    filter?: FilterOptions
-  ) => {
-    callOrderApi(async () => {
-      const data = await orderApi.getOrders(page, limit, keyword, sortBy, filter)
-      if (data) {
-        setOrders(data?.data?.orders || [])
-        setPagination(data?.data?.pagination || {})
-        const param = getNavigateParams(limit, keyword, sortBy, filter)
-        navigate(
-          `?page=${data?.data?.pagination?.currentPage}${param.limit}${param.keyword}${param.sortBy}${param.filter}`
-        )
-      } else {
-        // toast
-      }
-    })
-  }
-  const fetchOrderWithCurrentParams = () => {
-    fetchOrders(
-      queryParamPage,
-      queryParamLimit,
-      currentKeyword ? currentKeyword : undefined,
-      currentSortBy,
-      currentFilter
-    )
-  }
-  const onChangePage = (page: number, size: number) => {
-    if (page !== pagination.currentPage || size != pagination.limit) {
-      if (inputKeyword !== currentKeyword) {
-        setInputKeyword(currentKeyword)
-      }
-      fetchOrders(page, size, currentKeyword, currentSortBy, currentFilter)
-    }
-  }
-  const handleSearchOrder = () => {
-    if (inputKeyword !== currentKeyword) {
-      setCurrentKeyword(inputKeyword)
-      fetchOrders(1, pagination.limit, inputKeyword ? inputKeyword : undefined, currentSortBy, currentFilter)
-    }
-  }
-  const handleChangeSortOption = (value: SortOptions) => {
-    setCurrentSortBy(value)
-    fetchOrders(queryParamPage, queryParamLimit, currentKeyword, value, currentFilter)
-  }
-  const handleChangeFilterOption = (value: FilterOptions) => {
-    setCurrentFilter(value)
-    fetchOrders(queryParamPage, queryParamLimit, currentKeyword, currentSortBy, value)
-  }
-  useEffect(() => {
-    fetchOrderWithCurrentParams()
-  }, [])
-  return (
-    <div className='w-full h-full'>
-      <div className='h-full w-[95%] m-auto '>
-        <div>
-          <div className='flex flex-row w-[95%] h-fit justify-between m-auto mt-6'>
-            <div className='text-black-light text-2xl font-bold leading-tight text-left'>Order List</div>
-            <div>
-              <div className='flex items-end gap-4'>
-                <div className='flex items-center gap-2'>
-                  <span className='flex items-center text-lg text-black'>Filter:</span>
-                  <Dropdown
-                    value={currentFilter}
-                    onChange={(e) => handleChangeFilterOption(e.value)}
-                    options={filterOptions}
-                    optionLabel='label'
-                    className='w-40'
-                  />
-                </div>
-                <div className='flex items-center gap-2'>
-                  <span className='flex items-center text-lg text-black'>Sort:</span>
-                  <Dropdown
-                    optionLabel='label'
-                    optionValue='value'
-                    value={currentSortBy}
-                    onChange={(e) => handleChangeSortOption(e.value)}
-                    options={sortOptions}
-                    valueTemplate={selectedOptionTemplate}
-                    itemTemplate={sortOptionTemplate}
-                    className='w-36'
-                  />
-                </div>
-                <div className={`relative flex items-center gap-2 `}>
-                  <Button
-                    size='small'
-                    className='absolute w-10 h-10 p-2 text-2xl text-gray-500 ring-0 left-1'
-                    type='submit'
-                    rounded
-                    text
-                    icon={icons.search}
-                    onClick={handleSearchOrder}
-                  ></Button>
-                  <InputText
-                    placeholder='Order ID, Product name,...'
-                    size={'small'}
-                    className={` pl-10  w-80 `}
-                    onChange={(e) => setInputKeyword(e.target.value)}
-                    onKeyDown={(e) => {
-                      if (e.key === 'Enter') {
-                        handleSearchOrder()
-                      }
-                    }}
-                  />
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div className=' bg-white mt-4 rounded-lg'>
-          <div>
-            <DataTable
-              onMouseDownCapture={(e) => {
-                e.stopPropagation()
-              }}
-              value={orders}
-              tableStyle={{ minWidth: '50rem' }}
-              loading={callOrderApiLoading}
-              className='rounded-lg overflow-hidden'
-            >
-              <Column
-                alignHeader={'center'}
-                header='ID'
-                body={(data: IOrderReturn) => (
-                  <div
-                    className={`text-left font-semibold !select-text ${data.deletedAt && 'text-red-500 line-through opacity-50'}`}
-                  >
-                    {data.id}
-                  </div>
-                )}
-              />
-              <Column
-                alignHeader={'center'}
-                header='User'
-                body={(data: IOrderReturn) => (
-                  <div className={`text-left ${data.deletedAt && 'line-through opacity-50'}`}>{data?.user?.email}</div>
-                )}
-              />
-              <Column
-                alignHeader={'center'}
-                header='Product'
-                body={(data: IOrderReturn) => (
-                  <div className={`text-left flex items-start gap-2 ${data.deletedAt && 'opacity-50'}`}>
-                    <img className='w-20 h-auto' src={data?.products[0]?.imgUrl}></img>
-                    <div className='flex flex-col '>
-                      <div className='max-w-48 line-clamp-2 font-semibold'>{data?.products[0]?.name}</div>
-                      <div>
-                        {data?.products[0]?.colorName} / {data?.products[0]?.size}
-                      </div>
-                      <div className={`font-semibold text-red-500 ${data.deletedAt && 'line-through'}`}>
-                        {data?.products[0]?.price.toLocaleString('de-De')}đ
-                      </div>
-                    </div>{' '}
-                  </div>
-                )}
-              />
-              <Column
-                alignHeader={'right'}
-                header='Total price'
-                body={(data: IOrderReturn) => (
-                  <div className={`text-red-500 text-right ${data.deletedAt && 'line-through opacity-50'}`}>
-                    {data.totalPrice.toLocaleString('de-DE')}đ
-                  </div>
-                )}
-              />
-              <Column
-                alignHeader={'center'}
-                header='Order status'
-                body={(data: IOrderReturn) => (
-                  <div className={`${data.deletedAt && 'line-through opacity-50'}`}>
-                    {
-                      <div
-                        className={` text-center ${
-                          [OrderStatus.Delivered, OrderStatus.Confirmed, OrderStatus.Delivering].includes(
-                            data?.orderStatus
-                          )
-                            ? 'text-green-500'
-                            : data?.orderStatus === OrderStatus.Pending
-                              ? 'text-slate-500'
-                              : 'text-red-500'
-                        }`}
-                      >
-                        {getOrderStatusByEnum(data?.orderStatus)}
-                      </div>
-                    }
-                  </div>
-                )}
-              />
-              <Column
-                alignHeader={'center'}
-                header='Payment status'
-                body={(data: IOrderReturn) => (
-                  <div className={`${data.deletedAt && 'line-through opacity-50'}`}>
-                    {
-                      <div
-                        className={` text-center ${
-                          data.paymentStatus === PaymentStatus.Paid ? 'text-green-500' : 'text-red-500'
-                        }`}
-                      >
-                        {getPaymentStatusByEnum(data.paymentStatus)}
-                      </div>
-                    }
-                  </div>
-                )}
-              />
-              <Column
-                alignHeader={'center'}
-                header='Method'
-                body={(data: IOrderReturn) => (
-                  <div className={`${data.deletedAt && 'line-through opacity-50'}`}>
-                    {
-                      <div
-                        className={` text-center ${data.paymentMethod === PaymentMethod.Stripe ? 'text-blue-500' : 'text-amber-500'}`}
-                      >
-                        {data.paymentMethod === PaymentMethod.Stripe ? 'Stripe' : 'COD'}
-                      </div>
-                    }
-                  </div>
-                )}
-              />
-              <Column
-                alignHeader={'center'}
-                header='Paid at'
-                body={(data: IOrderReturn) =>
-                  data.paidAt ? (
-                    <div
-                      className={`text-center flex flex-col items-center ${data.deletedAt && 'line-through opacity-50'}`}
-                    >
-                      <div className='text-center'>{ConvertTimeString(data.paidAt)} </div>
-                      <div className='text-center'>{ConvertDateString(data.paidAt)}</div>
-                    </div>
-                  ) : (
-                    <div className={` text-center text-red-500`}>No data</div>
-                  )
-                }
-              />
-              <Column
-                alignHeader={'center'}
-                header='Created at'
-                body={(data: IOrderReturn) => (
-                  <div
-                    className={`text-center flex flex-col items-center ${data.deletedAt && 'line-through opacity-50'}`}
-                  >
-                    <div className='text-center'>{ConvertTimeString(data.createdAt)} </div>
-                    <div className='text-center'>{ConvertDateString(data.createdAt)}</div>
-                  </div>
-                )}
-              />
-              <Column
-                alignHeader={'center'}
-                header='Action'
-                body={(data: IOrderReturn) => (
-                  <div className='flex items-center justify-center'>
-                    <Button
-                      type='button'
-                      rounded
-                      text
-                      className='w-10 h-10 p-0 flex items-center justify-center ring-0'
-                      onClick={(e) => {
-                        setSelectedOrder(data)
-                        if (dropdownControlRef.current) {
-                          dropdownControlRef.current.toggle(e)
-                        }
-                      }}
-                    >
-                      <div className='text-2xl text-center'>{icons.list}</div>
-                    </Button>
-                    {dropdownControlRef && (
-                      <OverlayPanel className='drop-shadow-md p-2' ref={dropdownControlRef}>
-                        {selectedOrder &&
-                          (selectedOrder.deletedAt !== null ? (
-                            <Button
-                              text
-                              className='ring-0 px-3 py-1 flex items-center gap-2 text-green-500 select-none hover:cursor-pointer w-full'
-                              onClick={(e) => {
-                                handleRestoreOrder()
-                                if (dropdownControlRef.current) {
-                                  dropdownControlRef.current.toggle(e)
-                                }
-                              }}
-                            >
-                              {icons.restore} Restore
-                            </Button>
-                          ) : (
-                            <Button
-                              text
-                              className='ring-0 px-3 py-1 flex items-center gap-2 text-red-500 select-none hover:cursor-pointer w-full'
-                              onClick={deleteConfirm}
-                            >
-                              {icons.delete} Delete
-                            </Button>
-                          ))}
-
-                        {selectedOrder &&
-                          (selectedOrder.deletedAt !== null ? (
-                            <Button
-                              text
-                              className='ring-0 px-3 py-1 text-primary select-none hover:cursor-pointer w-full  flex items-center gap-2'
-                              disabled={selectedOrder.deletedAt !== null}
-                            >
-                              {icons.search} Detail
-                            </Button>
-                          ) : (
-                            <Link to={`/admin/order-detail/${selectedOrder.id}`}>
-                              <Button
-                                text
-                                className='ring-0 px-3 py-1 text-primary select-none hover:cursor-pointer w-full  flex items-center gap-2'
-                                disabled={selectedOrder.deletedAt !== null}
-                              >
-                                {icons.search} Detail
-                              </Button>
-                            </Link>
-                          ))}
-                      </OverlayPanel>
-                    )}
-                  </div>
-                )}
-              />
-            </DataTable>
-          </div>
-          <div>
-            <Paginator
-              rows={pagination?.limit}
-              first={((pagination?.currentPage || 1) - 1) * (pagination?.limit || 5)}
-              totalRecords={pagination?.totalOrders}
-              rowsPerPageOptions={[pagination?.limit && pagination?.limit < 10 ? pagination.limit : 5, 10, 25, 50]}
-              onPageChange={(e) => onChangePage(e.page + 1, e.rows)}
-            />
-          </div>
-        </div>
-      </div>
-      <ConfirmDialog
-        acceptClassName='bg-red-500 opacity-95 ring-0 border-none hover:opacity-100 hover:drop-shadow-md'
-        rejectClassName='text-slate-500 bg-transparent ring-0 border-none hover:opacity-90'
-      />
-    </div>
-  )
-}
